+++ src/devSVG.cpp
@@ -53,6 +53,7 @@ public:
   bool standalone;
   bool fix_text_size;
   double scaling;
+  bool keep_valid;
   const std::string file;
   cpp11::list system_aliases;
   cpp11::list user_aliases;
@@ -61,13 +62,14 @@ public:
 
   SVGDesc(SvgStreamPtr stream_, bool standalone_, cpp11::list aliases_,
           const std::string webfonts_, const std::string& file_, cpp11::strings ids_,
-          bool fix_text_size_, double scaling_):
+          bool fix_text_size_, double scaling_, bool keep_valid_):
       stream(stream_),
       pageno(0),
       clipx0(0), clipx1(0), clipy0(0), clipy1(0),
       standalone(standalone_),
       fix_text_size(fix_text_size_),
       scaling(scaling_),
+      keep_valid(keep_valid_),
       file(file_),
       system_aliases(cpp11::as_cpp<cpp11::list>(aliases_["system"])),
       user_aliases(cpp11::as_cpp<cpp11::list>(aliases_["user"])),
@@ -78,7 +80,7 @@ public:
   void nextFile() {
     stream->finish(false);
     if (stream->is_file_stream()) {
-      SvgStreamPtr newStream(new SvgStreamFile(file, pageno + 1));
+      SvgStreamPtr newStream(new SvgStreamFile(file, pageno + 1, keep_valid));
       stream = newStream;
     }
     clipid.clear();
@@ -909,7 +911,7 @@ pDevDesc svg_driver_new(SvgStreamPtr stream, int bg, double width,
                         bool standalone, cpp11::list& aliases,
                         const std::string& webfonts,
                         const std::string& file, cpp11::strings id,
-                        bool fix_text_size, double scaling) {
+                        bool fix_text_size, double scaling, bool keep_valid) {
 
   pDevDesc dd = (DevDesc*) calloc(1, sizeof(DevDesc));
   if (dd == NULL)
@@ -987,14 +989,15 @@ pDevDesc svg_driver_new(SvgStreamPtr stream, int bg, double width,
 #endif
 
   dd->deviceSpecific = new SVGDesc(stream, standalone, aliases, webfonts, file,
-                                   id, fix_text_size, scaling);
+                                   id, fix_text_size, scaling, keep_valid);
   return dd;
 }
 
 void makeDevice(SvgStreamPtr stream, std::string bg_, double width, double height,
                 double pointsize, bool standalone, cpp11::list& aliases,
                 const std::string& webfonts, const std::string& file,
-                cpp11::strings id, bool fix_text_size, double scaling) {
+                cpp11::strings id, bool fix_text_size, double scaling,
+                bool keep_valid) {
 
   int bg = R_GE_str2col(bg_.c_str());
 
@@ -1003,7 +1006,7 @@ void makeDevice(SvgStreamPtr stream, std::string bg_, double width, double heigh
   BEGIN_SUSPEND_INTERRUPTS {
     pDevDesc dev = svg_driver_new(stream, bg, width, height, pointsize,
                                   standalone, aliases, webfonts, file, id,
-                                  fix_text_size, scaling);
+                                  fix_text_size, scaling, keep_valid);
     if (dev == NULL)
       cpp11::stop("Failed to start SVG device");
 
@@ -1018,11 +1021,11 @@ void makeDevice(SvgStreamPtr stream, std::string bg_, double width, double heigh
 bool svglite_(std::string file, std::string bg, double width, double height,
               double pointsize, bool standalone, cpp11::list aliases,
               std::string webfonts, cpp11::strings id, bool fix_text_size,
-              double scaling) {
+              double scaling, bool keep_valid) {
 
-  SvgStreamPtr stream(new SvgStreamFile(file, 1));
+  SvgStreamPtr stream(new SvgStreamFile(file, 1, keep_valid));
   makeDevice(stream, bg, width, height, pointsize, standalone, aliases, webfonts,
-             file, id, fix_text_size, scaling);
+             file, id, fix_text_size, scaling, keep_valid);
 
   return true;
 }
@@ -1036,7 +1039,7 @@ cpp11::external_pointer<std::stringstream> svgstring_(cpp11::environment env, st
 
   SvgStreamPtr stream(new SvgStreamString(env));
   makeDevice(stream, bg, width, height, pointsize, standalone, aliases, webfonts,
-             "", id, fix_text_size, scaling);
+             "", id, fix_text_size, scaling, true);
 
   SvgStreamString* strstream = static_cast<SvgStreamString*>(stream.get());