+++ NEWS.md
@@ -35,6 +35,9 @@
 
 * Add scaling argument to devices to control line-width and text scaling (#115)
 
+* svg files that are being written are now only valid at all times if `keep_valid`
+  is set to `TRUE` in the `svglite()` call.
+
 
 # svglite 1.2.3